+++ controllers/argocd/TOBEREMOVED.go
@@ -1230,3 +1230,96 @@ func shortenHostname(hostname string) (string, error) {
 	}
 	return resultHostname, nil
 }
+
+// newService returns a new Service for the given ArgoCD instance.
+func newService(cr *argoproj.ArgoCD) *corev1.Service {
+	return &corev1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      cr.Name,
+			Namespace: cr.Namespace,
+			Labels:    argoutil.LabelsForCluster(cr),
+		},
+	}
+}
+
+// newServiceWithName returns a new Service instance for the given ArgoCD using the given name.
+func newServiceWithName(name string, component string, cr *argoproj.ArgoCD) *corev1.Service {
+	svc := newService(cr)
+	svc.ObjectMeta.Name = name
+
+	lbls := svc.ObjectMeta.Labels
+	lbls[common.ArgoCDKeyName] = name
+	lbls[common.ArgoCDKeyComponent] = component
+	svc.ObjectMeta.Labels = lbls
+
+	return svc
+}
+
+// newServiceWithSuffix returns a new Service instance for the given ArgoCD using the given suffix.
+func newServiceWithSuffix(suffix string, component string, cr *argoproj.ArgoCD) *corev1.Service {
+	return newServiceWithName(fmt.Sprintf("%s-%s", cr.Name, suffix), component, cr)
+}
+
+// reconcileGrafanaService will ensure that the Service for Grafana is present.
+func (r *ReconcileArgoCD) reconcileGrafanaService(cr *argoproj.ArgoCD) error {
+	svc := newServiceWithSuffix("grafana", "grafana", cr)
+	if argoutil.IsObjectFound(r.Client, cr.Namespace, svc.Name, svc) {
+		if !cr.Spec.Grafana.Enabled {
+			// Service exists but enabled flag has been set to false, delete the Service
+			return r.Client.Delete(context.TODO(), svc)
+		}
+		log.Info(grafanaDeprecatedWarning)
+		return nil // Service found, do nothing
+	}
+
+	if !cr.Spec.Grafana.Enabled {
+		return nil // Grafana not enabled, do nothing.
+	}
+
+	log.Info(grafanaDeprecatedWarning)
+	return nil
+}
+
+// reconcileServices will ensure that all Services are present for the given ArgoCD.
+func (r *ReconcileArgoCD) reconcileServices(cr *argoproj.ArgoCD) error {
+
+	if err := r.reconcileDexService(cr); err != nil {
+		log.Error(err, "error reconciling dex service")
+	}
+
+	err := r.reconcileGrafanaService(cr)
+	if err != nil {
+		return err
+	}
+
+	err = r.reconcileMetricsService(cr)
+	if err != nil {
+		return err
+	}
+
+	err = r.reconcileRedisHAServices(cr)
+	if err != nil {
+		return err
+	}
+
+	err = r.reconcileRedisService(cr)
+	if err != nil {
+		return err
+	}
+
+	err = r.reconcileRepoService(cr)
+	if err != nil {
+		return err
+	}
+
+	err = r.reconcileServerMetricsService(cr)
+	if err != nil {
+		return err
+	}
+
+	err = r.reconcileServerService(cr)
+	if err != nil {
+		return err
+	}
+	return nil
+}
