--- controllers/argocd/hpa.go
@@ -1,63 +1,0 @@
-// Copyright 2019 ArgoCD Operator Developers
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package argocd
-
-import (
-	autoscaling "k8s.io/api/autoscaling/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-	argoproj "github.com/argoproj-labs/argocd-operator/api/v1beta1"
-	"github.com/argoproj-labs/argocd-operator/common"
-	"github.com/argoproj-labs/argocd-operator/pkg/argoutil"
-)
-
-var (
-	maxReplicas int32 = 3
-	minReplicas int32 = 1
-	tcup        int32 = 50
-)
-
-func newHorizontalPodAutoscaler(cr *argoproj.ArgoCD) *autoscaling.HorizontalPodAutoscaler {
-	return &autoscaling.HorizontalPodAutoscaler{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      cr.Name,
-			Namespace: cr.Namespace,
-			Labels:    argoutil.LabelsForCluster(cr),
-		},
-	}
-}
-
-func newHorizontalPodAutoscalerWithName(name string, cr *argoproj.ArgoCD) *autoscaling.HorizontalPodAutoscaler {
-	hpa := newHorizontalPodAutoscaler(cr)
-	hpa.ObjectMeta.Name = name
-
-	lbls := hpa.ObjectMeta.Labels
-	lbls[common.ArgoCDKeyName] = name
-	hpa.ObjectMeta.Labels = lbls
-
-	return hpa
-}
-
-func newHorizontalPodAutoscalerWithSuffix(suffix string, cr *argoproj.ArgoCD) *autoscaling.HorizontalPodAutoscaler {
-	return newHorizontalPodAutoscalerWithName(nameWithSuffix(suffix, cr), cr)
-}
-
-// reconcileAutoscalers will ensure that all HorizontalPodAutoscalers are present for the given ArgoCD.
-func (r *ReconcileArgoCD) reconcileAutoscalers(cr *argoproj.ArgoCD) error {
-	if err := r.reconcileServerHPA(cr); err != nil {
-		return err
-	}
-	return nil
-}
