+++ controllers/argocd/route.go
@@ -16,14 +16,12 @@ package argocd
 
 import (
 	"context"
-	"fmt"
 
 	routev1 "github.com/openshift/api/route/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
 	argoproj "github.com/argoproj-labs/argocd-operator/api/v1beta1"
-	"github.com/argoproj-labs/argocd-operator/common"
 	"github.com/argoproj-labs/argocd-operator/pkg/argoutil"
 )
 
@@ -124,88 +122,4 @@ func (r *ReconcileArgoCD) reconcilePrometheusRoute(cr *argoproj.ArgoCD) error {
 		return err
 	}
 	return r.Client.Create(context.TODO(), route)
-}
-
-// reconcileApplicationSetControllerWebhookRoute will ensure that the ArgoCD Server Route is present.
-func (r *ReconcileArgoCD) reconcileApplicationSetControllerWebhookRoute(cr *argoproj.ArgoCD) error {
-	name := fmt.Sprintf("%s-%s", common.ApplicationSetServiceNameSuffix, "webhook")
-	route := newRouteWithSuffix(name, cr)
-	found := argoutil.IsObjectFound(r.Client, cr.Namespace, route.Name, route)
-	if found {
-		if cr.Spec.ApplicationSet == nil || !cr.Spec.ApplicationSet.WebhookServer.Route.Enabled {
-			// Route exists but enabled flag has been set to false, delete the Route
-			return r.Client.Delete(context.TODO(), route)
-		}
-	}
-
-	if cr.Spec.ApplicationSet == nil || !cr.Spec.ApplicationSet.WebhookServer.Route.Enabled {
-		return nil // Route not enabled, move along...
-	}
-
-	// Allow override of the Annotations for the Route.
-	if len(cr.Spec.ApplicationSet.WebhookServer.Route.Annotations) > 0 {
-		route.Annotations = cr.Spec.Server.Route.Annotations
-	}
-
-	// Allow override of the Labels for the Route.
-	if len(cr.Spec.ApplicationSet.WebhookServer.Route.Labels) > 0 {
-		labels := route.Labels
-		for key, val := range cr.Spec.Server.Route.Labels {
-			labels[key] = val
-		}
-		route.Labels = labels
-	}
-
-	// Allow override of the Host for the Route.
-	if len(cr.Spec.Server.Host) > 0 {
-		route.Spec.Host = cr.Spec.ApplicationSet.WebhookServer.Host
-	}
-
-	hostname, err := shortenHostname(route.Spec.Host)
-	if err != nil {
-		return err
-	}
-
-	route.Spec.Host = hostname
-
-	if cr.Spec.Server.Insecure {
-		// Disable TLS and rely on the cluster certificate.
-		route.Spec.Port = &routev1.RoutePort{
-			TargetPort: intstr.FromString("webhook"),
-		}
-		route.Spec.TLS = &routev1.TLSConfig{
-			InsecureEdgeTerminationPolicy: routev1.InsecureEdgeTerminationPolicyRedirect,
-			Termination:                   routev1.TLSTerminationEdge,
-		}
-	} else {
-		// Server is using TLS configure passthrough.
-		route.Spec.Port = &routev1.RoutePort{
-			TargetPort: intstr.FromString("webhook"),
-		}
-		route.Spec.TLS = &routev1.TLSConfig{
-			InsecureEdgeTerminationPolicy: routev1.InsecureEdgeTerminationPolicyRedirect,
-			Termination:                   routev1.TLSTerminationPassthrough,
-		}
-	}
-
-	// Allow override of TLS options for the Route
-	if cr.Spec.Server.Route.TLS != nil {
-		route.Spec.TLS = cr.Spec.Server.Route.TLS
-	}
-
-	route.Spec.To.Kind = "Service"
-	route.Spec.To.Name = nameWithSuffix(common.ApplicationSetServiceNameSuffix, cr)
-
-	// Allow override of the WildcardPolicy for the Route
-	if cr.Spec.Server.Route.WildcardPolicy != nil && len(*cr.Spec.Server.Route.WildcardPolicy) > 0 {
-		route.Spec.WildcardPolicy = *cr.Spec.Server.Route.WildcardPolicy
-	}
-
-	if err := controllerutil.SetControllerReference(cr, route, r.Scheme); err != nil {
-		return err
-	}
-	if !found {
-		return r.Client.Create(context.TODO(), route)
-	}
-	return r.Client.Update(context.TODO(), route)
 }
