--- controllers/argocd/service.go
@@ -1,148 +1,0 @@
-// Copyright 2019 ArgoCD Operator Developers
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package argocd
-
-import (
-	"context"
-	"fmt"
-
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/intstr"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-
-	argoproj "github.com/argoproj-labs/argocd-operator/api/v1beta1"
-	"github.com/argoproj-labs/argocd-operator/common"
-	"github.com/argoproj-labs/argocd-operator/pkg/argoutil"
-)
-
-// newService returns a new Service for the given ArgoCD instance.
-func newService(cr *argoproj.ArgoCD) *corev1.Service {
-	return &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      cr.Name,
-			Namespace: cr.Namespace,
-			Labels:    argoutil.LabelsForCluster(cr),
-		},
-	}
-}
-
-// newServiceWithName returns a new Service instance for the given ArgoCD using the given name.
-func newServiceWithName(name string, component string, cr *argoproj.ArgoCD) *corev1.Service {
-	svc := newService(cr)
-	svc.ObjectMeta.Name = name
-
-	lbls := svc.ObjectMeta.Labels
-	lbls[common.ArgoCDKeyName] = name
-	lbls[common.ArgoCDKeyComponent] = component
-	svc.ObjectMeta.Labels = lbls
-
-	return svc
-}
-
-// newServiceWithSuffix returns a new Service instance for the given ArgoCD using the given suffix.
-func newServiceWithSuffix(suffix string, component string, cr *argoproj.ArgoCD) *corev1.Service {
-	return newServiceWithName(fmt.Sprintf("%s-%s", cr.Name, suffix), component, cr)
-}
-
-// reconcileGrafanaService will ensure that the Service for Grafana is present.
-func (r *ReconcileArgoCD) reconcileGrafanaService(cr *argoproj.ArgoCD) error {
-	svc := newServiceWithSuffix("grafana", "grafana", cr)
-	if argoutil.IsObjectFound(r.Client, cr.Namespace, svc.Name, svc) {
-		if !cr.Spec.Grafana.Enabled {
-			// Service exists but enabled flag has been set to false, delete the Service
-			return r.Client.Delete(context.TODO(), svc)
-		}
-		log.Info(grafanaDeprecatedWarning)
-		return nil // Service found, do nothing
-	}
-
-	if !cr.Spec.Grafana.Enabled {
-		return nil // Grafana not enabled, do nothing.
-	}
-
-	log.Info(grafanaDeprecatedWarning)
-	return nil
-}
-
-// reconcileServerMetricsService will ensure that the Service for the Argo CD server metrics is present.
-func (r *ReconcileArgoCD) reconcileServerMetricsService(cr *argoproj.ArgoCD) error {
-	svc := newServiceWithSuffix("server-metrics", "server", cr)
-	if argoutil.IsObjectFound(r.Client, cr.Namespace, svc.Name, svc) {
-		return nil // Service found, do nothing
-	}
-
-	svc.Spec.Selector = map[string]string{
-		common.ArgoCDKeyName: nameWithSuffix("server", cr),
-	}
-
-	svc.Spec.Ports = []corev1.ServicePort{
-		{
-			Name:       "metrics",
-			Port:       common.ServerMetricsPort,
-			Protocol:   corev1.ProtocolTCP,
-			TargetPort: intstr.FromInt(common.ServerMetricsPort),
-		},
-	}
-
-	if err := controllerutil.SetControllerReference(cr, svc, r.Scheme); err != nil {
-		return err
-	}
-	return r.Client.Create(context.TODO(), svc)
-}
-
-// reconcileServices will ensure that all Services are present for the given ArgoCD.
-func (r *ReconcileArgoCD) reconcileServices(cr *argoproj.ArgoCD) error {
-
-	if err := r.reconcileDexService(cr); err != nil {
-		log.Error(err, "error reconciling dex service")
-	}
-
-	err := r.reconcileGrafanaService(cr)
-	if err != nil {
-		return err
-	}
-
-	err = r.reconcileMetricsService(cr)
-	if err != nil {
-		return err
-	}
-
-	err = r.reconcileRedisHAServices(cr)
-	if err != nil {
-		return err
-	}
-
-	err = r.reconcileRedisService(cr)
-	if err != nil {
-		return err
-	}
-
-	err = r.reconcileRepoService(cr)
-	if err != nil {
-		return err
-	}
-
-	err = r.reconcileServerMetricsService(cr)
-	if err != nil {
-		return err
-	}
-
-	err = r.reconcileServerService(cr)
-	if err != nil {
-		return err
-	}
-	return nil
-}
